Fix broken item% totals when powerups are cvar-replaced
ZDoom gets confused if a Stop is used to run off of the end of a
state, since Lost Souls and Pain Elementals do that after they die, so
it can't remove them from the end-tally since they still count that
way.

So for the cvar-based powerups, we need to explicitly tell ZDoom that
we want to remove them from the map so it knows to drop them from the
item% totals.